requirements.txt
geopandas
shapely
fiona
pyproj
rasterio
pdfplumber
pytesseract
pillow
opencv-python
laspy
pdal
ezdxf
jsonschema
